+++ crates/keychain/src/hw_backed/mod.rs
@@ -7,9 +7,9 @@ mod secure_enclave;
 pub use secure_enclave::{delete_key, retrieve_key, store_key};
 
 #[cfg(target_os = "windows")]
-mod windows_hello;
+mod windows_tpm;
 #[cfg(target_os = "windows")]
-pub use windows_hello::{delete_key, retrieve_key, store_key};
+pub use windows_tpm::{delete_key, retrieve_key, store_key};
 
 #[cfg(target_os = "linux")]
 mod linux_tpm;